refactor Nitrox (no. 2)
