+++ src/cli/index.ts
@@ -6,9 +6,10 @@ import { setTimeout } from "timers/promises";
 import { packagePrompt, install, configPrompt} from "~/utils/prompts.js";
 import { PackageManager } from "~/utils/getPackageManager.js";
 import { turboCLI } from "./turbo.js";
-
-import { nitroxCLI } from "./nitrox.js";
+import { nitroxPrompt } from "./nitrox.js";
 import { Task, tasks } from "~/utils/task.js";
+import { intro } from "~/utils/prompts/intro.js";
+import { outro } from "~/utils/prompts/outro.js";
 
 const defaultOptions: cliResults = {
     flags: {
@@ -33,8 +34,7 @@ export const runCLI = async (packageManager: PackageManager, flags: cliFlags): P
         return cliResults;
     }
 
-    prompt.intro(gradient.atlas("HUCLI"));
-    await setTimeout(1000)
+    await intro();
     prompt.note("Let's get you configured.", gradient.atlas("Step 1:"))
 
     const config = await configPrompt(packageManager, cliResults.flags)
@@ -87,11 +87,10 @@ export const runCLI = async (packageManager: PackageManager, flags: cliFlags): P
 
     if (cliResults.flags.nitrox) {
         await setTimeout(1000)
-        await nitroxCLI(packageManager, cliResults.flags)
+        await nitroxPrompt(packageManager, cliResults.flags)
     }
-    await setTimeout(1000);
-
-    prompt.outro(`You are now ready to ${gradient.atlas("Hit The Ground Running")}!`);
+    
+    await outro();
 
     return cliResults;
 }
